elf is parsing now.

+++ include/stdio.h
@@ -3,6 +3,6 @@
 
 int printf(const char *format, ... );
 int scanf(const char *format, ...);
-char* itoa( int num ); 
+char* itoa( unsigned long int num ); 
 void putc( char c); 
 #endif
